ml_engine/requirements.txt
flask
pandas
scikit-learn
numpy
python-dotenv
requests
gunicorn
joblib
psycopg2-binary
sqlalchemy
scipy
